requirements.txt
aiosqlite~=0.19.0
aiogram~=3.0.0RC1
python-dotenv~=1.0.0
python-telegram-bot~=20.4